chore(content-preview): PR feedback
