+++ src/api/Versions.js
@@ -5,7 +5,7 @@
  */
 
 import getProp from 'lodash/get';
-import { VERSIONS_FIELDS_TO_FETCH } from '../utils/fields';
+import { FILE_VERSIONS_FIELDS_TO_FETCH } from '../utils/fields';
 import OffsetBasedAPI from './OffsetBasedAPI';
 import {
     DEFAULT_FETCH_END,
@@ -196,7 +196,7 @@ class Versions extends OffsetBasedAPI {
         errorCallback: ElementsErrorCallback,
         offset: number = DEFAULT_FETCH_START,
         limit: number = DEFAULT_FETCH_END,
-        fields: Array<string> = VERSIONS_FIELDS_TO_FETCH,
+        fields: Array<string> = FILE_VERSIONS_FIELDS_TO_FETCH,
         shouldFetchAll: boolean = true,
     ): void {
         this.errorCode = ERROR_CODE_FETCH_VERSIONS;
