+++ src/elements/content-preview/preview-header/Header.scss
@@ -36,7 +36,7 @@
     }
 }
 
-.bcpr-header--dark {
+.bcpr-header--basic {
     .bp-header {
         background: $threes;
         border-bottom: 0;
