+++ src/elements/content-preview/preview-header/index.js
@@ -1,2 +1,2 @@
 // @flow
-export { default } from './PreviewHeaderComponent';
+export { default } from './PreviewHeader';
