+++ src/utils/__tests__/fields-test.js
@@ -3,7 +3,7 @@ import {
     PREVIEW_FIELDS_TO_FETCH,
     SIDEBAR_FIELDS_TO_FETCH,
     TASKS_FIELDS_TO_FETCH,
-    VERSIONS_FIELDS_TO_FETCH,
+    FILE_VERSIONS_FIELDS_TO_FETCH,
     TASK_ASSIGNMENTS_FIELDS_TO_FETCH,
     COMMENTS_FIELDS_TO_FETCH,
     findMissingProperties,
@@ -126,10 +126,12 @@ describe('util/fields', () => {
     });
 
     test('should fetch correct version fields', () => {
-        expect(VERSIONS_FIELDS_TO_FETCH).toEqual([
+        expect(FILE_VERSIONS_FIELDS_TO_FETCH).toEqual([
+            FIELD_EXTENSION,
             FIELD_CREATED_AT,
             FIELD_MODIFIED_AT,
             FIELD_MODIFIED_BY,
+            FIELD_NAME,
             FIELD_SIZE,
             FIELD_TRASHED_AT,
             FIELD_VERSION_NUMBER,
